requirements.txt
fastapi
uvicorn[standard]

langchain
langchain-core
langgraph

langchain-ollama
pydantic
python-dotenv

typing-extensions

openai>=1.40.0
